requirements.txt
langchain>=0.1.0
langchain-community>=0.0.0
langchain-text-splitters>=0.0.0
pypdf>=3.0.0
openai>=0.27.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
unstructured>=0.9.0
sentence-transformers>=2.2.0